move GetHTTPOption to pkg/networking and add comment
