+++ pkg/networking/util_test.go
@@ -1,4 +1,4 @@
-package reconciler
+package networking
 
 import (
 	"errors"
